requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
jupyter>=1.0.0
python-dotenv>=1.0.0

# CLI and utilities
click>=8.1.0

# Data export and reporting
matplotlib>=3.7.0
reportlab>=4.0.0
openpyxl>=3.1.0

# Optional 3D visualization
pyvista>=0.40.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
